+++ plugins/woocommerce/client/admin/client/analytics/settings/historical-data/period-selector.js
@@ -65,6 +65,7 @@ function HistoricalDataPeriodSelector( {
 		<div className="woocommerce-settings-historical-data__columns">
 			<div className="woocommerce-settings-historical-data__column">
 				<SelectControl
+					__next40pxDefaultSize
 					label={ __( 'Import historical data', 'woocommerce' ) }
 					value={ value.label }
 					disabled={ disabled }