+++ plugins/woocommerce/src/Blocks/BlockTypes/Cart.php
@@ -244,6 +244,31 @@ class Cart extends AbstractBlock {
 		$this->asset_data_registry->add( 'isBlockTheme', wp_is_block_theme() );
 		$this->asset_data_registry->add( 'shippingMethodsExist', CartCheckoutUtils::shipping_methods_exist() > 0 );
 
+		$is_block_editor = $this->is_block_editor();
+
+		// Check `current_user_can` so we can show notices about incompatible extensions in the front-end to admins too.
+		if ( ( $is_block_editor || current_user_can( 'manage_woocommerce' ) ) && ! $this->asset_data_registry->exists( 'incompatibleExtensions' ) ) {
+			if ( class_exists( '\Automattic\WooCommerce\Utilities\FeaturesUtil' ) && function_exists( 'get_plugins' ) ) {
+				$declared_extensions     = \Automattic\WooCommerce\Utilities\FeaturesUtil::get_compatible_plugins_for_feature( 'cart_checkout_blocks' );
+				$all_plugins             = \get_plugins();
+				$incompatible_extensions = array_reduce(
+					$declared_extensions['incompatible'],
+					function ( array $acc, $item ) use ( $all_plugins ) {
+						$plugin      = $all_plugins[ $item ] ?? null;
+						$plugin_id   = $plugin['TextDomain'] ?? dirname( $item );
+						$plugin_name = $plugin['Name'] ?? $plugin_id;
+						$acc[]       = [
+							'id'    => $plugin_id,
+							'title' => $plugin_name,
+						];
+						return $acc;
+					},
+					[]
+				);
+				$this->asset_data_registry->add( 'incompatibleExtensions', $incompatible_extensions );
+			}
+		}
+
 		// Hydrate the following data depending on admin or frontend context.
 		if ( ! is_admin() && ! WC()->is_rest_api_request() ) {
 			$this->asset_data_registry->hydrate_api_request( '/wc/store/v1/cart' );